fix bug,has to hardcode make sure normal,texcoord has position,bcz maybe next program use these,and our vao reuse the attribute location info,and not inited again

+++ test/src/shaders/shadow/depth.frag
@@ -1,7 +1,8 @@
 #version 300 es
-precision mediump float;
+precision highp float;
 
 
-void main(){
+void main()
+{   
 
 }
